B3.28: canonical main signature + hint-arity follow-up
Reverse engineers reading dac's C output flagged `main` rendering as
`int64_t main(void)` even though every C compiler wants `int main`.
This batch adds a curated entry-point catalogue (`main`, `wmain`,
`WinMain`) and threads its C-canonical spellings through a new
SynthesizedSignature override channel in the C backend so `main`
reads as `int main(void)` (or `int main(int argc, char **argv)` when
rdi/rsi are live).

The same channel covers the hint-arity follow-up: a `[[function]]`
hint that declares more arg slots than the convention pass observed
now mints synthetic `RegisterArg` entries past the observed prefix,
absorbing the residue-shelf "hint argument synthesis" item.

The canonical overlay declines when observed > catalogue max so a
PE main reading `rcx, rdx, r8, r9` (CRT-side wrapper) keeps its body
references intact — the IR stays ground truth (I-1).

10 new tests: 7 in `dac-cli::lift::tests`, 3 in
`dac-backend-c::lower::tests`. 6 catalogue tests in
`dac-knowledge::canonical::tests`. Four goldens refresh; PE main left
untouched per the declining-overlay rule.

Closes B3.28, FR-12, FR-20, FR-21, I-1, I-3.

+++ tests/golden/hello-elf-o1-c/source.c
@@ -62,7 +62,7 @@ int64_t _init(int64_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t ar
 /* stack_locals: 1 */
 /* struct_layouts: pointer=0 stack=0 */
 /* switch_tables: 0 */
-int64_t main(void) {
+int main(void) {
     int64_t v0 = 0LL;
     void * v1 = ((void *)(0LL));
     int64_t v2 = 0LL;
@@ -76,7 +76,7 @@ int64_t main(void) {
     *((int64_t *)(((int64_t)(v1)))) = v2;
     v9 = ((long long (*)(long long, long long, long long, long long, long long, long long))write)(1LL, 8196LL, 6LL, v6, v7, v8);
     v11 = (*((int64_t *)(((int64_t)(v1)))));
-    return 42LL;
+    return ((int)(42LL));
 }
 
 /* dac-recovered function */
